Adds FPS debug service
Adds an FPS debug service to display the current frames per second.

Renames GameSystems directory to GameServices for clarity and consistency.

Moves debug input logic to DebugInputService.cs and BeatSyncDebugger to BeatSyncDebugService.cs to keep it aligned with the new folder.

Adds a toggle to show FPS to the InputActions